common: finish reorganize/rename in common
