+++ packages/common/src/enums.ts
@@ -1,0 +1,39 @@
+export enum ChainId {
+  Mainnet = 1,
+  Ropsten = 3,
+  Rinkeby = 4,
+  Kovan = 42,
+  Goerli = 5,
+  Sepolia = 11155111,
+}
+
+export enum HardforkName {
+  Chainstart = 'chainstart',
+  Homestead = 'homestead',
+  Dao = 'dao',
+  TangerineWhistle = 'tangerineWhistle',
+  SpuriousDragon = 'spuriousDragon',
+  Byzantium = 'byzantium',
+  Constantinople = 'constantinople',
+  Petersburg = 'petersburg',
+  Istanbul = 'istanbul',
+  MuirGlacier = 'muirGlacier',
+  Berlin = 'berlin',
+  London = 'london',
+  ArrowGlacier = 'arrowGlacier',
+  Shanghai = 'shanghai',
+  MergeForkIdTransition = 'mergeForkIdTransition',
+  Merge = 'merge',
+}
+
+export enum ConsensusType {
+  ProofOfStake = 'pos',
+  ProofOfWork = 'pow',
+  ProofOfAuthority = 'poa',
+}
+
+export enum ConsensusAlgorithm {
+  Ethash = 'ethash',
+  Clique = 'clique',
+  Casper = 'casper',
+}
