+++ packages/common/tests/chains.spec.ts
@@ -1,5 +1,5 @@
 import tape from 'tape'
-import Common, { Chain, ConsensusAlgorithm, ConsensusType, Hardfork } from '../src/'
+import Common, { ChainId, ConsensusAlgorithm, ConsensusType, HardforkName } from '../src/'
 
 tape('[Common/Chains]: Initialization / Chain params', function (t: tape.Test) {
   t.test('Should initialize with chain provided', function (st: tape.Test) {
@@ -7,7 +7,7 @@ tape('[Common/Chains]: Initialization / Chain params', function (t: tape.Test) {
     st.equal(c.chainName(), 'mainnet', 'should initialize with chain name')
     st.equal(c.chainId(), BigInt(1), 'should return correct chain Id')
     st.equal(c.networkId(), BigInt(1), 'should return correct network Id')
-    st.equal(c.hardfork(), Hardfork.London, 'should set hardfork to current default hardfork')
+    st.equal(c.hardfork(), HardforkName.London, 'should set hardfork to current default hardfork')
     st.equal(
       c.hardfork(),
       c.DEFAULT_HARDFORK,
@@ -21,11 +21,11 @@ tape('[Common/Chains]: Initialization / Chain params', function (t: tape.Test) {
   })
 
   t.test('Should initialize with chain provided by Chain enum', function (st: tape.Test) {
-    const c = new Common({ chain: Chain.Mainnet })
+    const c = new Common({ chain: ChainId.Mainnet })
     st.equal(c.chainName(), 'mainnet', 'should initialize with chain name')
     st.equal(c.chainId(), BigInt(1), 'should return correct chain Id')
     st.equal(c.networkId(), BigInt(1), 'should return correct network Id')
-    st.equal(c.hardfork(), Hardfork.London, 'should set hardfork to current default hardfork')
+    st.equal(c.hardfork(), HardforkName.London, 'should set hardfork to current default hardfork')
     st.equal(
       c.hardfork(),
       c.DEFAULT_HARDFORK,
@@ -45,7 +45,7 @@ tape('[Common/Chains]: Initialization / Chain params', function (t: tape.Test) {
   t.test(
     'Should initialize with chain and hardfork provided by Chain and Hardfork enums',
     function (st: tape.Test) {
-      const c = new Common({ chain: Chain.Mainnet, hardfork: Hardfork.Byzantium })
+      const c = new Common({ chain: ChainId.Mainnet, hardfork: HardforkName.Byzantium })
       st.equal(c.hardfork(), 'byzantium', 'should return correct hardfork name')
 
       st.end()
