+++ packages/common/tests/hardforks.spec.ts
@@ -1,28 +1,28 @@
 import tape from 'tape'
-import Common, { Chain, ConsensusAlgorithm, ConsensusType, Hardfork } from '../src'
+import Common, { ChainId, ConsensusAlgorithm, ConsensusType, HardforkName } from '../src'
 
 tape('[Common]: Hardfork logic', function (t: tape.Test) {
   t.test('Hardfork access', function (st: tape.Test) {
     const supportedHardforks = [
-      Hardfork.Chainstart,
-      Hardfork.Homestead,
-      Hardfork.Dao,
-      Hardfork.Chainstart,
-      Hardfork.SpuriousDragon,
-      Hardfork.Byzantium,
-      Hardfork.Constantinople,
-      Hardfork.Petersburg,
-      Hardfork.Istanbul,
-      Hardfork.Berlin,
-      Hardfork.London,
-      Hardfork.ArrowGlacier,
-      Hardfork.Shanghai,
-      Hardfork.Merge,
+      HardforkName.Chainstart,
+      HardforkName.Homestead,
+      HardforkName.Dao,
+      HardforkName.Chainstart,
+      HardforkName.SpuriousDragon,
+      HardforkName.Byzantium,
+      HardforkName.Constantinople,
+      HardforkName.Petersburg,
+      HardforkName.Istanbul,
+      HardforkName.Berlin,
+      HardforkName.London,
+      HardforkName.ArrowGlacier,
+      HardforkName.Shanghai,
+      HardforkName.Merge,
     ]
     let c
 
     for (const hardfork of supportedHardforks) {
-      c = new Common({ chain: Chain.Mainnet, hardfork: hardfork })
+      c = new Common({ chain: ChainId.Mainnet, hardfork: hardfork })
       st.equal(c.hardfork(), hardfork, hardfork)
     }
 
@@ -30,58 +30,58 @@ tape('[Common]: Hardfork logic', function (t: tape.Test) {
   })
 
   t.test('getHardforkByBlockNumber() / setHardforkByBlockNumber()', function (st: tape.Test) {
-    let c = new Common({ chain: Chain.Mainnet })
+    let c = new Common({ chain: ChainId.Mainnet })
     let msg = 'should get HF correctly'
 
-    st.equal(c.getHardforkByBlockNumber(0), Hardfork.Chainstart, msg)
-    st.equal(c.getHardforkByBlockNumber(1149999), Hardfork.Chainstart, msg)
-    st.equal(c.getHardforkByBlockNumber(1150000), Hardfork.Homestead, msg)
-    st.equal(c.getHardforkByBlockNumber(1400000), Hardfork.Homestead, msg)
-    st.equal(c.getHardforkByBlockNumber(9200000), Hardfork.MuirGlacier, msg)
-    st.equal(c.getHardforkByBlockNumber(12244000), Hardfork.Berlin, msg)
-    st.equal(c.getHardforkByBlockNumber(12965000), Hardfork.London, msg)
-    st.equal(c.getHardforkByBlockNumber(13773000), Hardfork.ArrowGlacier, msg)
-    st.equal(c.getHardforkByBlockNumber(999999999999), Hardfork.ArrowGlacier, msg)
+    st.equal(c.getHardforkByBlockNumber(0), HardforkName.Chainstart, msg)
+    st.equal(c.getHardforkByBlockNumber(1149999), HardforkName.Chainstart, msg)
+    st.equal(c.getHardforkByBlockNumber(1150000), HardforkName.Homestead, msg)
+    st.equal(c.getHardforkByBlockNumber(1400000), HardforkName.Homestead, msg)
+    st.equal(c.getHardforkByBlockNumber(9200000), HardforkName.MuirGlacier, msg)
+    st.equal(c.getHardforkByBlockNumber(12244000), HardforkName.Berlin, msg)
+    st.equal(c.getHardforkByBlockNumber(12965000), HardforkName.London, msg)
+    st.equal(c.getHardforkByBlockNumber(13773000), HardforkName.ArrowGlacier, msg)
+    st.equal(c.getHardforkByBlockNumber(999999999999), HardforkName.ArrowGlacier, msg)
 
     msg = 'should set HF correctly'
 
-    st.equal(c.setHardforkByBlockNumber(0), Hardfork.Chainstart, msg)
-    st.equal(c.setHardforkByBlockNumber(1149999), Hardfork.Chainstart, msg)
-    st.equal(c.setHardforkByBlockNumber(1150000), Hardfork.Homestead, msg)
-    st.equal(c.setHardforkByBlockNumber(1400000), Hardfork.Homestead, msg)
-    st.equal(c.setHardforkByBlockNumber(12244000), Hardfork.Berlin, msg)
-    st.equal(c.setHardforkByBlockNumber(12965000), Hardfork.London, msg)
-    st.equal(c.setHardforkByBlockNumber(13773000), Hardfork.ArrowGlacier, msg)
-    st.equal(c.setHardforkByBlockNumber(999999999999), Hardfork.ArrowGlacier, msg)
+    st.equal(c.setHardforkByBlockNumber(0), HardforkName.Chainstart, msg)
+    st.equal(c.setHardforkByBlockNumber(1149999), HardforkName.Chainstart, msg)
+    st.equal(c.setHardforkByBlockNumber(1150000), HardforkName.Homestead, msg)
+    st.equal(c.setHardforkByBlockNumber(1400000), HardforkName.Homestead, msg)
+    st.equal(c.setHardforkByBlockNumber(12244000), HardforkName.Berlin, msg)
+    st.equal(c.setHardforkByBlockNumber(12965000), HardforkName.London, msg)
+    st.equal(c.setHardforkByBlockNumber(13773000), HardforkName.ArrowGlacier, msg)
+    st.equal(c.setHardforkByBlockNumber(999999999999), HardforkName.ArrowGlacier, msg)
 
-    c = new Common({ chain: Chain.Ropsten })
+    c = new Common({ chain: ChainId.Ropsten })
     st.equal(c.setHardforkByBlockNumber(0), 'tangerineWhistle', msg)
 
     st.end()
   })
 
   t.test('setHardfork(): hardforkChanged event', function (st) {
-    const c = new Common({ chain: Chain.Mainnet, hardfork: Hardfork.Istanbul })
+    const c = new Common({ chain: ChainId.Mainnet, hardfork: HardforkName.Istanbul })
     c.on('hardforkChanged', (hardfork: string) => {
-      st.equal(hardfork, Hardfork.Byzantium, 'should send correct hardforkChanged event')
+      st.equal(hardfork, HardforkName.Byzantium, 'should send correct hardforkChanged event')
       st.end()
     })
-    c.setHardfork(Hardfork.Byzantium)
+    c.setHardfork(HardforkName.Byzantium)
   })
 
   t.test('hardforkBlock()', function (st: tape.Test) {
-    let c = new Common({ chain: Chain.Ropsten })
+    let c = new Common({ chain: ChainId.Ropsten })
     let msg = 'should return the correct HF change block for byzantium (provided)'
-    st.equal(c.hardforkBlock(Hardfork.Byzantium)!, BigInt(1700000), msg)
+    st.equal(c.hardforkBlock(HardforkName.Byzantium)!, BigInt(1700000), msg)
 
     msg = 'should return null if HF does not exist on chain'
     st.equal(c.hardforkBlock('thisHardforkDoesNotExist'), null, msg)
 
-    c = new Common({ chain: Chain.Ropsten, hardfork: Hardfork.Byzantium })
+    c = new Common({ chain: ChainId.Ropsten, hardfork: HardforkName.Byzantium })
     msg = 'should return the correct HF change block for byzantium (set)'
     st.equal(c.hardforkBlock()!, BigInt(1700000), msg)
 
-    c = new Common({ chain: Chain.Ropsten, hardfork: Hardfork.Istanbul })
+    c = new Common({ chain: ChainId.Ropsten, hardfork: HardforkName.Istanbul })
     msg = 'should return the correct HF change block for istanbul (set)'
     st.equal(c.hardforkBlock()!, BigInt(6485846), msg)
 
@@ -89,14 +89,14 @@ tape('[Common]: Hardfork logic', function (t: tape.Test) {
   })
 
   t.test('isHardforkBlock()', function (st: tape.Test) {
-    let c = new Common({ chain: Chain.Ropsten })
+    let c = new Common({ chain: ChainId.Ropsten })
     let msg = 'should return true for HF change block for byzantium (provided)'
-    st.equal(c.isHardforkBlock(1700000, Hardfork.Byzantium), true, msg)
+    st.equal(c.isHardforkBlock(1700000, HardforkName.Byzantium), true, msg)
 
     msg = 'should return false for another block for byzantium (provided)'
-    st.equal(c.isHardforkBlock(1700001, Hardfork.Byzantium), false, msg)
+    st.equal(c.isHardforkBlock(1700001, HardforkName.Byzantium), false, msg)
 
-    c = new Common({ chain: Chain.Ropsten, hardfork: Hardfork.Byzantium })
+    c = new Common({ chain: ChainId.Ropsten, hardfork: HardforkName.Byzantium })
     msg = 'should return true for HF change block for byzantium (set)'
     st.equal(c.isHardforkBlock(1700000), true, msg)
 
@@ -107,7 +107,7 @@ tape('[Common]: Hardfork logic', function (t: tape.Test) {
   })
 
   t.test('nextHardforkBlock()', function (st: tape.Test) {
-    let c = new Common({ chain: Chain.Rinkeby, hardfork: Hardfork.Chainstart })
+    let c = new Common({ chain: ChainId.Rinkeby, hardfork: HardforkName.Chainstart })
     let msg =
       'should work with HF set / return correct next HF block for chainstart (rinkeby: chainstart -> homestead)'
     st.equal(c.nextHardforkBlock()!, BigInt(1), msg)
@@ -117,21 +117,21 @@ tape('[Common]: Hardfork logic', function (t: tape.Test) {
     st.equal(c.nextHardforkBlock('homestead')!, BigInt(2), msg)
 
     msg = 'should return correct next HF (rinkeby: byzantium -> constantinople)'
-    st.equal(c.nextHardforkBlock(Hardfork.Byzantium)!, BigInt(3660663), msg)
+    st.equal(c.nextHardforkBlock(HardforkName.Byzantium)!, BigInt(3660663), msg)
 
     msg = 'should return null if next HF is not available (rinkeby: london -> shanghai)'
-    st.equal(c.nextHardforkBlock(Hardfork.London), null, msg)
+    st.equal(c.nextHardforkBlock(HardforkName.London), null, msg)
 
     msg =
       'should work correctly along the need to skip several forks (ropsten: chainstart -> (homestead) -> (dao) -> (tangerineWhistle) -> spuriousDragon)'
-    c = new Common({ chain: Chain.Ropsten, hardfork: Hardfork.Chainstart })
+    c = new Common({ chain: ChainId.Ropsten, hardfork: HardforkName.Chainstart })
     st.equal(c.nextHardforkBlock()!, BigInt(10), msg)
 
     st.end()
   })
 
   t.test('isNextHardforkBlock()', function (st: tape.Test) {
-    const c = new Common({ chain: Chain.Rinkeby, hardfork: Hardfork.Chainstart })
+    const c = new Common({ chain: ChainId.Rinkeby, hardfork: HardforkName.Chainstart })
     let msg =
       'should work with HF set / return true for correct next HF block for chainstart (rinkeby: chainstart -> homestead)'
     st.equal(c.isNextHardforkBlock(1), true, msg)
@@ -141,29 +141,29 @@ tape('[Common]: Hardfork logic', function (t: tape.Test) {
     st.equal(c.isNextHardforkBlock(2, 'homestead'), true, msg)
 
     msg = 'should return true for correct next HF (rinkeby: byzantium -> constantinople)'
-    st.equal(c.isNextHardforkBlock(3660663, Hardfork.Byzantium), true, msg)
+    st.equal(c.isNextHardforkBlock(3660663, HardforkName.Byzantium), true, msg)
 
     msg = 'should return false for a block number too low (rinkeby: byzantium -> constantinople)'
-    st.equal(c.isNextHardforkBlock(124, Hardfork.Byzantium), false, msg)
+    st.equal(c.isNextHardforkBlock(124, HardforkName.Byzantium), false, msg)
 
     msg = 'should return false for a block number too high (rinkeby: byzantium -> constantinople)'
-    st.equal(c.isNextHardforkBlock(605948938, Hardfork.Byzantium), false, msg)
+    st.equal(c.isNextHardforkBlock(605948938, HardforkName.Byzantium), false, msg)
 
     st.end()
   })
 
   t.test('hardforkIsActiveOnBlock() / activeOnBlock()', function (st: tape.Test) {
-    let c = new Common({ chain: Chain.Ropsten })
+    let c = new Common({ chain: ChainId.Ropsten })
     let msg = 'Ropsten, byzantium (provided), 1700000 -> true'
-    st.equal(c.hardforkIsActiveOnBlock(Hardfork.Byzantium, 1700000), true, msg)
+    st.equal(c.hardforkIsActiveOnBlock(HardforkName.Byzantium, 1700000), true, msg)
 
     msg = 'Ropsten, byzantium (provided), 1700005 -> true'
-    st.equal(c.hardforkIsActiveOnBlock(Hardfork.Byzantium, 1700005), true, msg)
+    st.equal(c.hardforkIsActiveOnBlock(HardforkName.Byzantium, 1700005), true, msg)
 
     msg = 'Ropsten, byzantium (provided), 1699999 -> false'
-    st.equal(c.hardforkIsActiveOnBlock(Hardfork.Byzantium, 1699999), false, msg)
+    st.equal(c.hardforkIsActiveOnBlock(HardforkName.Byzantium, 1699999), false, msg)
 
-    c = new Common({ chain: Chain.Ropsten, hardfork: Hardfork.Byzantium })
+    c = new Common({ chain: ChainId.Ropsten, hardfork: HardforkName.Byzantium })
     msg = 'Ropsten, byzantium (set), 1700000 -> true'
     st.equal(c.hardforkIsActiveOnBlock(null, 1700000), true, msg)
 
@@ -180,62 +180,66 @@ tape('[Common]: Hardfork logic', function (t: tape.Test) {
   })
 
   t.test('hardforkBlock()', function (st: tape.Test) {
-    const c = new Common({ chain: Chain.Mainnet })
+    const c = new Common({ chain: ChainId.Mainnet })
 
     let msg = 'should return correct value'
-    st.equal(c.hardforkBlock(Hardfork.Berlin)!, BigInt(12244000), msg)
-    st.equal(c.hardforkBlock(Hardfork.Berlin)!, BigInt(12244000), msg)
+    st.equal(c.hardforkBlock(HardforkName.Berlin)!, BigInt(12244000), msg)
+    st.equal(c.hardforkBlock(HardforkName.Berlin)!, BigInt(12244000), msg)
 
     msg = 'should return null for unscheduled hardfork'
     // developer note: when Shanghai is set,
     // update this test to next unscheduled hardfork.
-    st.equal(c.hardforkBlock(Hardfork.Shanghai), null, msg)
-    st.equal(c.hardforkBlock(Hardfork.Shanghai), null, msg)
-    st.equal(c.nextHardforkBlock(Hardfork.Shanghai), null, msg)
+    st.equal(c.hardforkBlock(HardforkName.Shanghai), null, msg)
+    st.equal(c.hardforkBlock(HardforkName.Shanghai), null, msg)
+    st.equal(c.nextHardforkBlock(HardforkName.Shanghai), null, msg)
 
     st.end()
   })
 
   t.test('hardforkGteHardfork()', function (st: tape.Test) {
-    let c = new Common({ chain: Chain.Ropsten })
+    let c = new Common({ chain: ChainId.Ropsten })
     let msg = 'Ropsten, constantinople >= byzantium (provided) -> true'
-    st.equal(c.hardforkGteHardfork(Hardfork.Constantinople, Hardfork.Byzantium), true, msg)
+    st.equal(c.hardforkGteHardfork(HardforkName.Constantinople, HardforkName.Byzantium), true, msg)
 
     msg = 'Ropsten, dao >= chainstart (provided) -> false'
-    st.equal(c.hardforkGteHardfork(Hardfork.Dao, Hardfork.Chainstart), false, msg)
+    st.equal(c.hardforkGteHardfork(HardforkName.Dao, HardforkName.Chainstart), false, msg)
 
     msg = 'Ropsten, byzantium >= byzantium (provided) -> true'
-    st.equal(c.hardforkGteHardfork(Hardfork.Byzantium, Hardfork.Byzantium), true, msg)
+    st.equal(c.hardforkGteHardfork(HardforkName.Byzantium, HardforkName.Byzantium), true, msg)
 
     msg = 'Ropsten, spuriousDragon >= byzantium (provided) -> false'
-    st.equal(c.hardforkGteHardfork(Hardfork.SpuriousDragon, Hardfork.Byzantium), false, msg)
+    st.equal(c.hardforkGteHardfork(HardforkName.SpuriousDragon, HardforkName.Byzantium), false, msg)
 
-    c = new Common({ chain: Chain.Ropsten, hardfork: Hardfork.Byzantium })
+    c = new Common({ chain: ChainId.Ropsten, hardfork: HardforkName.Byzantium })
     msg = 'Ropsten, byzantium (set) >= spuriousDragon -> true'
-    st.equal(c.hardforkGteHardfork(null, Hardfork.SpuriousDragon), true, msg)
+    st.equal(c.hardforkGteHardfork(null, HardforkName.SpuriousDragon), true, msg)
 
     msg = 'Ropsten, byzantium (set) >= spuriousDragon -> true (alias function)'
-    st.equal(c.gteHardfork(Hardfork.SpuriousDragon), true, msg)
+    st.equal(c.gteHardfork(HardforkName.SpuriousDragon), true, msg)
 
     msg = 'Ropsten, byzantium (set) >= byzantium -> true'
-    st.equal(c.hardforkGteHardfork(null, Hardfork.Byzantium), true, msg)
+    st.equal(c.hardforkGteHardfork(null, HardforkName.Byzantium), true, msg)
 
     msg = 'Ropsten, byzantium (set) >= constantinople -> false'
-    st.equal(c.hardforkGteHardfork(null, Hardfork.Constantinople), false, msg)
+    st.equal(c.hardforkGteHardfork(null, HardforkName.Constantinople), false, msg)
 
     st.end()
   })
 
   t.test('hardforkGteHardfork() ropsten', function (st: tape.Test) {
-    const c = new Common({ chain: Chain.Ropsten })
+    const c = new Common({ chain: ChainId.Ropsten })
     const msg = 'ropsten, spuriousDragon >= muirGlacier (provided) -> false'
-    st.equal(c.hardforkGteHardfork(Hardfork.SpuriousDragon, Hardfork.MuirGlacier), false, msg)
+    st.equal(
+      c.hardforkGteHardfork(HardforkName.SpuriousDragon, HardforkName.MuirGlacier),
+      false,
+      msg
+    )
 
     st.end()
   })
 
   t.test('_calcForkHash()', function (st: tape.Test) {
-    let c = new Common({ chain: Chain.Mainnet })
+    let c = new Common({ chain: ChainId.Mainnet })
     let msg = 'should calc correctly for chainstart (only genesis)'
     st.equal(c._calcForkHash('chainstart'), '0xfc64ec04', msg)
 
@@ -243,7 +247,7 @@ tape('[Common]: Hardfork logic', function (t: tape.Test) {
     st.equal(c._calcForkHash('homestead'), '0x97c2c34c', msg)
 
     msg = 'should calc correctly for in-between applied HF'
-    st.equal(c._calcForkHash(Hardfork.Byzantium), '0xa00bc324', msg)
+    st.equal(c._calcForkHash(HardforkName.Byzantium), '0xa00bc324', msg)
 
     const chains = ['mainnet', 'ropsten', 'rinkeby', 'goerli', 'kovan']
 
@@ -262,16 +266,16 @@ tape('[Common]: Hardfork logic', function (t: tape.Test) {
   })
 
   t.test('forkHash()', function (st: tape.Test) {
-    let c = new Common({ chain: Chain.Mainnet, hardfork: Hardfork.Byzantium })
+    let c = new Common({ chain: ChainId.Mainnet, hardfork: HardforkName.Byzantium })
     let msg = 'should provide correct forkHash for HF set'
     st.equal(c.forkHash(), '0xa00bc324', msg)
 
     msg = 'should provide correct forkHash for HF provided'
-    st.equal(c.forkHash(Hardfork.SpuriousDragon), '0x3edd5b10', msg)
+    st.equal(c.forkHash(HardforkName.SpuriousDragon), '0x3edd5b10', msg)
 
-    c = new Common({ chain: Chain.Kovan })
+    c = new Common({ chain: ChainId.Kovan })
     let f = () => {
-      c.forkHash(Hardfork.Merge)
+      c.forkHash(HardforkName.Merge)
     }
     msg = 'should throw when called on non-applied or future HF'
     st.throws(f, /No fork hash calculation possible/, msg)
@@ -286,11 +290,11 @@ tape('[Common]: Hardfork logic', function (t: tape.Test) {
   })
 
   t.test('hardforkForForkHash()', function (st: tape.Test) {
-    const c = new Common({ chain: Chain.Mainnet })
+    const c = new Common({ chain: ChainId.Mainnet })
 
     let msg = 'should return the correct HF array for a matching forkHash'
     const res = c.hardforkForForkHash('0x3edd5b10')!
-    st.equal(res.name, Hardfork.SpuriousDragon, msg)
+    st.equal(res.name, HardforkName.SpuriousDragon, msg)
 
     msg = 'should return null for a forkHash not matching any HF'
     st.equal(c.hardforkForForkHash('0x12345'), null, msg)
@@ -299,7 +303,7 @@ tape('[Common]: Hardfork logic', function (t: tape.Test) {
   })
 
   t.test('HF consensus updates', function (st: tape.Test) {
-    let c = new Common({ chain: Chain.Goerli, hardfork: Hardfork.Byzantium })
+    let c = new Common({ chain: ChainId.Goerli, hardfork: HardforkName.Byzantium })
     st.equal(
       c.consensusType(),
       ConsensusType.ProofOfAuthority,
@@ -316,7 +320,7 @@ tape('[Common]: Hardfork logic', function (t: tape.Test) {
       'should provide the correct initial chain consensus configuration'
     )
 
-    c = new Common({ chain: Chain.Goerli, hardfork: Hardfork.Merge })
+    c = new Common({ chain: ChainId.Goerli, hardfork: HardforkName.Merge })
     st.equal(
       c.consensusType(),
       ConsensusType.ProofOfStake,
